save: add a new option for skipping packages required by test files
Current godep save tries to save packages which are required by test
related files in gopath. If test files are not required (-t isn't
passed), these packages are not needed to be stored under Godeps. This
commit lets save command not to save such packages if a new option -T
is passed.